+++ apps/router-e2e/__e2e__/headless/app/styles/index.tsx
@@ -1,0 +1,9 @@
+import { Link } from 'expo-router';
+
+export default function StylesIndex() {
+  return (
+    <>
+      <Link href="./stylex">Style X</Link>
+    </>
+  );
+}