+++ apps/router-e2e/__e2e__/headless/app/styles/stylex.tsx
@@ -1,0 +1,5 @@
+import { Text } from 'react-native';
+
+export default function StyleXNative() {
+  return <Text>StyleX does not work on native</Text>;
+}