+++ apps/router-e2e/__e2e__/headless/stylex-compat.ts
@@ -1,0 +1,10 @@
+import * as stylex from '@stylexjs/stylex';
+
+export function stylexProps(...styles: any[]) {
+  let { className, style = [], ...props } = stylex.props(...styles);
+
+  if (!Array.isArray(style)) style = [style];
+  style.push({ $$css: true, [className]: className });
+
+  return { style, ...props };
+}