+++ apps/router-e2e/app.config.js
@@ -56,7 +56,7 @@ module.exports = {
             : process.env.E2E_ROUTER_ASYNC === 'false'
               ? false
               : process.env.E2E_ROUTER_ASYNC || false,
-        root: path.join('__e2e__', process.env.E2E_ROUTER_SRC ?? 'static-rendering', 'app'),
+        root: path.join('__e2e__', process.env.E2E_ROUTER_SRC ?? 'headless', 'app'),
         origin: 'https://smart-symbiote.netlify.app/',
       },
     ],