+++ packages/expo-router/src/exports.ts
@@ -33,4 +33,6 @@ export { SplashScreen } from './views/Splash';
 export { useNavigation } from './useNavigation';
 export { useFocusEffect } from './useFocusEffect';
 
+export * from './headless/Tabs';
+
 export type * from './types';